+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ENHANCEMENTS:
 * allow `initial` of `0` for `yandex_kubernetes_node_group` `auto_scale` policy.
 
+BUG FIXES:
+* alb: fix behavior of attribute `weight` in `http_backend` and `grpc_backend` and `stream_backend` in `yandex_alb_backend_group` resource
+
 ## 0.70.0 (January 17, 2022)
 FEATURES:
 * **New Resource:** `yandex_datatransfer_endpoint`