+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## 0.71.0 (Unreleased)
 ENHANCEMENTS:
 * allow `initial` of `0` for `yandex_kubernetes_node_group` `auto_scale` policy.
+* dataproc: now cluster version specified in create cluster request is treated as a semantic version prefix. So specified version `2.0` may lead to creation of cluster version `2.0.30` and this will not be considered a change in terraform state.
+* add `host_group_ids` attribute in `yandex_mdb_sqlserver_cluster` resource and data source
+
+BUG FIXES:
+* alb: fix behavior of attribute `weight` in `http_backend` and `grpc_backend` and `stream_backend` in `yandex_alb_backend_group` resource
+* datatransfer: make `yandex_datatransfer_transfer` importable in accordance with the documentation
 
 ## 0.70.0 (January 17, 2022)
 FEATURES: